+++ netfox/NFXStatisticsController.swift
@@ -5,16 +5,10 @@
 //  Copyright © 2015 kasketis. All rights reserved.
 //
 
-#if os(iOS)
-
 import Foundation
-import UIKit
 
 class NFXStatisticsController: NFXGenericController
 {
-    var scrollView: UIScrollView = UIScrollView()
-    var textLabel: UILabel = UILabel()
-
     var totalModels: Int = 0
 
     var successfulRequests: Int = 0
@@ -33,27 +27,7 @@ class NFXStatisticsController: NFXGenericController
     {
         super.viewDidLoad()
         
-        self.title = "Statistics"
-        
         generateStatics()
-        
-        self.scrollView = UIScrollView()
-        self.scrollView.frame = CGRectMake(0, 0, CGRectGetWidth(self.view.frame), CGRectGetHeight(self.view.frame))
-        self.scrollView.autoresizingMask = [.FlexibleWidth, .FlexibleHeight]
-        self.scrollView.autoresizesSubviews = true
-        self.scrollView.backgroundColor = UIColor.clearColor()
-        self.view.addSubview(self.scrollView)
-
-        self.textLabel = UILabel()
-        self.textLabel.frame = CGRectMake(20, 20, CGRectGetWidth(scrollView.frame) - 40, CGRectGetHeight(scrollView.frame) - 20);
-        self.textLabel.font = UIFont.NFXFont(13)
-        self.textLabel.textColor = UIColor.NFXGray44Color()
-        self.textLabel.numberOfLines = 0
-        self.textLabel.attributedText = getReportString()
-        self.textLabel.sizeToFit()
-        self.scrollView.addSubview(self.textLabel)
-        
-        self.scrollView.contentSize = CGSizeMake(CGRectGetWidth(scrollView.frame), CGRectGetMaxY(self.textLabel.frame))
         
         NSNotificationCenter.defaultCenter().addObserver(
             self,
@@ -155,10 +129,5 @@ class NFXStatisticsController: NFXGenericController
     {
         clearStatistics()
         generateStatics()
-        dispatch_async(dispatch_get_main_queue()) { () -> Void in
-            self.textLabel.attributedText = self.getReportString()
-        }
     }
 }
-
-#endif
